fix(gen): phase 3 pricing — 2× markup, provider-cost, snapshot, fractional
Kevin's pricing policy:
- 2× markup on provider cost (50% gross margin). NO 10×/15× markup.
- Pricing target: max(0.25 credits, provider_cost_usd × 2.0 / 0.01)
- Charge fractional credits (NUMERIC(18,6)); do NOT ceil per-component.
- Snapshot pricing at request start — admin updates don't change
  charges for in-flight or completed requests.
- Support cached input pricing + tool cost.

Replaces the Phase 2 actual-token charging shipped in PR #259 with
the proper provider-cost-based formula. PR #259's charge formula
was wrong: it read output_credit_rate as credits/1K-output-tokens,
but the DB values were set up as Phase 1 markup factors (e.g. 1.5×
pricing), so charges were off by an order of magnitude.

Files (6, +545/-2):
- supabase/migrations/20260803194600_pricing_policy_2x_pass_through.sql
  (new). Adds provider_input_usd_per_1m, provider_cached_input_usd_per_1m,
  provider_output_usd_per_1m, billing_multiplier, pricing_effective_at to
  generation_models. Adds pricing snapshot columns (input/output/cached
  rate per 1K, multiplier, minimum, credit value, effective_at) to
  generation_outputs. Backfills provider prices for currently enabled
  models (gpt-5.5 / gpt-4o / gpt-4o-mini / gpt-3.5-turbo / gpt-4.1 /
  gpt-4.1-mini / gpt-5.4-mini). Backfills minimum/credit_value/multiplier
  on existing generation_outputs rows that pre-date the snapshot.

- supabase/functions/generate-story/_generation_models.ts (+111).
  Added PricingSnapshot, GenerationUsage, PricingDefaults, DEFAULT_PRICING,
  snapshotPricing(model), computeActualChargeCredits(usage, pricing),
  computeMaxChargeCredits(usage, pricing). Kept Phase 1
  computeGenerationCreditCharge as legacy. New functions use
  provider-cost basis with 6-decimal precision (no ceil).

- supabase/functions/generate-story/_provider.ts (+6). Added
  cachedInputTokens and toolCostUsd to LLMResponse. OpenAI provider
  populates cachedInputTokens from
  json.usage?.input_tokens_details?.cached_tokens (responses API).

- supabase/functions/generate-story/index.ts (+34/-16). Added
  cachedInputTokens and toolCostUsd to LlmResult. Updated estimate
  branch, credit check, and post-flight charge to use
  snapshotPricing(selectedModel) + computeMaxChargeCredits /
  computeActualChargeCredits with a GenerationUsage struct.

- supabase/functions/generate-story/pricing_test.ts (new, +290).
  17 unit tests covering: input+output both charged, cached input rate,
  0.25 floor, fractional credits preserved (no ceil), 2× → 50% margin,
  long-context input + tool charges, snapshot derivation from model
  fields. Plus pre-flight max matches actual formula.

- CathedralOSApp/Services/GenerationRequestDTO.swift (+1/-1). Changed
  creditCostCharged from Int? to Double? (type + decode). The
  post-flight value is now fractional; the iOS display layer rounds
  for display but stores the raw Double.

DB seed (from migration):
  GPT-5.5: input $5.00/1M, cached $0.50/1M, output $30.00/1M
  GPT-4o: input $2.50/1M, cached $1.25/1M, output $10.00/1M
  GPT-4o-mini: input $0.15/1M, cached $0.075/1M, output $0.60/1M
  GPT-3.5-turbo: input $0.50/1M, cached $0.50/1M, output $1.50/1M
  GPT-4.1: input $2.00/1M, cached $0.50/1M, output $8.00/1M
  GPT-4.1-mini: input $0.40/1M, cached $0.10/1M, output $1.60/1M
  GPT-5.4-mini: input $0.40/1M, cached $0.10/1M, output $1.60/1M

  Derived credit rate per 1K = provider_usd_per_1m × multiplier / 10
  (since 1 credit = $0.01).

Example: gpt-5.5 with 1000 uncached + 1600 output tokens
  = (1000 × $5 + 1600 × $30) / 1M × 2.0 / $0.01 = 10.6 credits
  OpenAI cost: $0.053. Customer charge: $0.106. 50% margin.

Not validated on this host: WSL2 has no Deno toolchain, can't run tests.

Follow-up: drop legacy input_credit_rate/output_credit_rate/minimum_charge_credits
columns in a follow-up migration after iOS rolls out the fractional display.

+++ CathedralOSApp/Services/GenerationRequestDTO.swift
@@ -204,7 +204,7 @@ struct GenerationResponse: Codable {
     /// The number of credits available to the user (present on insufficient_credits error).
     let availableCredits: Int?
     /// The number of credits charged for this generation (present on success).
-    let creditCostCharged: Int?
+    let creditCostCharged: Double?
     /// The number of credits remaining after this generation (present on success).
     let remainingCredits: Int?
 
@@ -238,7 +238,7 @@ struct GenerationResponse: Codable {
         errorCode           = try c.decodeIfPresent(String.self, forKey: .errorCode)
         requiredCredits     = try c.decodeIfPresent(Int.self, forKey: .requiredCredits)
         availableCredits    = try c.decodeIfPresent(Int.self, forKey: .availableCredits)
-        creditCostCharged   = try c.decodeIfPresent(Int.self, forKey: .creditCostCharged)
+        creditCostCharged   = try c.decodeIfPresent(Double.self, forKey: .creditCostCharged)
         remainingCredits    = try c.decodeIfPresent(Int.self, forKey: .remainingCredits)
         retryAfterSeconds   = try c.decodeIfPresent(Int.self, forKey: .retryAfterSeconds)
         status              = try c.decode(String.self, forKey: .status)
